Font discovery and helpful comments
I believe that Elementary Statistics, 2nd edition uses the Stymie
font face. Stymie is a (Egyptian) Slab-Serif, created in 1931, related
to Memphis. Rockwell is a nice, slightly cleaner alternative, that has a
slightly more modern feel. While Stymie is an exact, or close to exact
match, Rockwell would be my choice to preserve and update the text's
style.